[refactor] Renamed ConnectionState to ConnectionStatemanager
